refactor: コメント長を規約に合わせ、設置送信をPlaceBlockProtocolSenderへ分離
