+++ moorestech_client/Assets/Scripts/Client.Game/InGame/BlockSystem/PlaceSystem/Util/PlaceBlockProtocolSender.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 969d9c05fd45c462a92fd79671e06fc7
